+++ test/scenarios/scenarios-dubbo-serialization/dubbo-serialization-jdk-test/src/main/java/org/apache/dubbo/test/serialization/jdk/JdkProvider.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.apache.dubbo.test.serialization.protostuff;
+package org.apache.dubbo.test.serialization.jdk;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
