+++ test/scenarios/scenarios-dubbo-serialization/dubbo-serialization-protostuff-test/src/main/java/org/apache/dubbo/test/serialization/jdk/ProtostuffProvider.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.apache.dubbo.test.serialization.protostuff;
+package org.apache.dubbo.test.serialization.jdk;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
